Update README with installation instructions. Move .gitignore to ignore .egg-info and pycharm artifacts. Fix circular dependencies
