+++ README.md
@@ -8,6 +8,8 @@ Frontend: svelte
 
 ## Get started
 
+We recommend using your favorite virtual environment while developing!
+
 Install the dependencies for svelte
 
 ```bash
@@ -15,11 +17,18 @@ cd client
 npm install
 ```
 
+Install the dependencies for niviz_rater and the package itself.
+```bash
+pip install -r requirements.txt
+pip install -e .
+```
+
 ## Run
 
 To run the bottle application, first initialize the database, 
 then run the Web Interface
 ```bash
+cd niviz_reader
 python app.py initialize_db
 python app.py runserver
 ```
